Replace ComposeViewController Setup. Minor fixes for sending messages
